Only flush rewrite rules on shutdown in a non switched context
If in the context of `switch_to_blog()`, the options of the new
site may pollute the global rewrite permastruct of the main site.
The shutdown action fires in the context of the main site and
may then populate rewrite rules on that main site based on the
permastruct of the new site. Phew.