Improve the UX of the snippet editor
Previously it wasn't clear for users that the editor was editable by
clicking on the fields. Now there is a big edit icon that guides you to
the smaller edit icons when you hover.

In the future it should be possible to edit the snippet editor fields
by clicking on the big edit icon and make the icon a button.