Optimisation in unit tests
Statistics:
Instead of testing for the full range, just testing the boundaries. This means that only 4 items need to be created instead of 100.

Frontend:
Instead of creating 22 items to test for 3 pages (with 10 items per page) just set items per page to 1 and test with 3 items.
Resetting post per page to the default in tearDown.