"Notification Center"
Yoast Alerts introduced
Moved all Notifications from inline to a separate page with dismissed and active alerts.

Notifications that are active should always be added to the Notification Center.
When they are not active they should not be added.

Not yet done for the PR:
Fix tests
Add new tests

JavaScript loading / unbind click events when an action has been started