Massively improve performance of passive voice
The passive voice makes extensive use of giant list of words that need
to be transformed to regexes. We were transforming these each time we
needed the regex. Now we cache them using lodash's memoize.

This can be much improved over time by shipping a data structure that is
already optimized instead of relying on the memory to save time.