Don't empty the text focus keyword field
We were emptying the focus keyword field when you switched to the
content tab, this meant that we had to check in multiple locations in
the code if the content might not be set. This PR changes this so the
focus keyword value stays in the text field. This makes the code easier
to reason about as well.

The reason this is possible is because we don't show the snippet preview
on the content tab anymore.