Create a UrlStructure object
A UrlStructure object represents the structure of with static and
variable parts. The static parts are just strings and can for example
represent the protocol and hostname part of the URL. A variable part can
be used together with data to build a complete URL.

This UrlStructure object can be instantiated at the creation of the App
to have more semantic information about how a URL is build on a specific
platform. When we have the UrlStructure we know which parts are static
and variable and we could make assessments based on this information.

The UrlStructure is also very useful when building the snippet editor,
it might be possible in the future to have URL fields be dynamically
generated based on this URL structure.