CS/XSS: always escape output - 32
The `title()` method retrieves the pre-escaped title from the `generate_title()` method. This pre-escaping is incorrect, however as not all theme authors escape correctly and the `esc_html()` method protects against double escaping, I think it's prudent to leave the pre-escaping in place.

https://github.com/Yoast/wordpress-seo/blob/c12c0926044dfc32dded185028637b8945e6fae6/frontend/class-frontend.php#L577

All the same, it makes sense to escape the title when it's being force replaced in the `flush_cache()` method.

Includes a minor efficiency fix within that same function by removing a duplicate function call.