CS/best practices: use `wp_rand()` instead of `rand()`
WP offers a `wp_rand()` function as an alternative to the PHP native `rand()` function. Using the WP variant is recommended as it is more secure and less predictable than the values generated by the PHP native functions.

Use of the PHP native `rand()` related functions is discouraged anyway as the values they generate are not not secure. Using functions like `random_int()`, `random_bytes()` and `openssl_random_pseudo_bytes()` are suggested as alternatives.

The WP wrapper function makes use of these alternatives under the hood whenever possible.

Refs:
* https://developer.wordpress.org/reference/functions/wp_rand/
* http://php.net/manual/en/function.rand.php