CS: clean up after latest merges - add missing translators comment
Regarding the second one (in `user-profile.php`):
PHP open/close tags for multi-line snippets should be on their own line, but this can cause layout issues when the output is directly embedded in certain HTML tags.
By creating a variable beforehand and passing that in, we create a stable result.

Includes some improved output escaping.