Bump version to not display the WHIP message on
As the ServeHappy dashboard message has not been implemented in the WordPress version we expected it to be in, we need to bump the version which hides the WHIP message.

The changelog for 4.9.6 has just been released and the ServeHappy code is not referenced in that one, bumping the version check to 4.9.7 for now.