Disallow partial selection of entities
We don't want to allow partial selection of entities because that
results in very unexpected behavior. Because we cannot distinguish
between mouse and keyboard input we have to compromise.

When moving into an entity with the keyboard now the whole entity will
be selected. When selecting half-way into an entity the whole entity
will be selected.