Fix bug where nl_NL would not contain all translations
But only sometimes. This was because it would be overwritten by its
formal counterpart. It seemed random because the last one to download
would be the file on disk.

This required a change in grunt-glotpress, because it didn't expose the
slug before. The slug has the formal/informal information. With this
change we can change the fileFormat. That makes sure that the files are
not overwritten.

Some additional changes are necessary to cleanup the languages folder.
All files with suffixes are removed. This is because we currently don't
support any formal languages. We can fix this later.