CS/I18n: Remove line breaks from translatable strings
This is a bit of an awkward one as it is unclear to me whether these phrases had line-breaks in them on purpose or accidentally.

If on purpose: the phrases were now often broken up mid-way in a sentence.
For translators, line-breaks in translatable strings -especially when at semi-random places - are difficult to work with as words have different lengths in different languages, let alone in different writing systems.

If there was a purpose to the line-breaks, it would be great to know the reason and we can figure what would be a better solution (if any).