CS: Move multi-line parameters out of function calls [10]
WPCS 1.1.0 introduced a stricter check on function call layouts.
Multi-line function calls now need to have each argument on a new line.
In a next iteration of this principle, it is expected that a sniff will be introduced to ban multi-line function call arguments.

See:
* https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards/releases/tag/1.1.0
* https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards/issues/ 1330

With this mind, function calls with multi-line parameters which are currently already causing errors to be thrown by PHPCS after the changes in WPCS 1.1.0, have been fixed - depending on the existing code - by:
* either changing multi-line function call arguments to single line;
* or by moving a multi-line function call arguments out of the function call and defining it as a variable before passing it to the function call.

This commit contains changes for the second variant.

As the `input` for the functions being tested, as well as the `expected output` is the same, these two test methods also contained duplicate information, making the code more difficult to maintain.

This code duplication has also been removed.

Note: a further abstraction of the test data to a class property could be considered, however as the data for the two functions involved is not 100% the same and the test class contains quite some test methods which don't use the same data, I felt that it would be more obvious what the tests were doing to keep the data as a class variable.