Travis/Composer: update YoastCS to 1.2.1 for the Premium plugin only
A bit hacki-di-hacky, but it works.

This adds a separate Composer script to allow both devs, as well as Travis, to test code in the Premium repo against YoastCS 1.2.1, even though the Free plugin is not ready to upgrade yet.

To still allows devs to easily switch between testing the Free repo against the old requirements and the Premium repo against the new ones, the update to YoastCS 1.2.1 for the Premium CS check is reverted after the CS checks have run.

Also takes note of the fact that `premium-check-cs` script contains **2** PHPCS runs.

The PHPCompatibility `testVersion` can not be set per directory (inherent to how config settings work in PHPCS), however the `testVersion` as set in a ruleset, can - since PHPCS 3.3.0 - be overruled from the command-line.
This allows us to test the PHP 5.2 compatible code with a `testVersion` of `5.2-`, which is the default in YoastCS, while we test the PHP 5.3+ compatible code (in the `cli` directory) with a `testVersion` of `5.3-`.

The same will be applied to the `check-cs` script once the Free plugin upgrades to the current version of YoastCS.
This also means that we won't need a slew of ruleset exclusions for directories with different PHP requirements.

The `fix-cs` script has not been adjusted as the `PHPCompatibility` standard does not contain fixers, so there won't be an issue there.

Lastly, as the Premium repo will have its own PHPCS ruleset, we can now remove Premium specific exclusions from the PHPCS ruleset used by the Free plugin.

Once this PR has been merged and Free merged back into Premium, you should be able to see the results of this change in the Travis PHP 7.3 build in Premium.

The proof of concept for this change has been tested extensively. Results for a (successful) test run on Travis using the same code as included in this PR, are available here: https://travis-ci.com/Yoast/wordpress-seo-premium/builds/95946057#L2403