Composer: sync with other config files
I've done a compare between the `composer.json` config files in (nearly) all plugin repos.

The only thing which appears to be needed for this repo is to tweak the `test` scripts for cross-platform compatibility.

I'd also like to propose removing the `check-cs-errors` script. Not sure what that was supposed to do and who ever used it, but whatever the intention was, there are simpler ways to achieve the same without touching the error/warning severity.