CS: don't use underscore prefixed property names [2]
Properties should not be prefixed with an `_` (to indicate visibility or for whatever other reason). This is an outdated PHP4-style habit.

Renaming the property and instances of it in child classes fixes this.

As the `Yoast_Model` class is set up as a parent class, this should be considered a BC-break.