[src] QA: call global functions using a fully qualified name
When PHP encounters an unqualified function call in a namespaced file, it will first try and find the function within the namespace and only when it can not find it in the namespace, fall-back to the global function.

Using fully qualified function names when calling a global function ensures that PHP will bypass the search within the namespace and call the global function directly.

It also ensures that optimal use is made of the PHP7 opcodes on VM level which have been implemented for a select group of PHP internal functions resulting in faster code execution.

Ref: https://www.php.net/manual/en/language.namespaces.fallback.php