Make the label and labelSingular dependent on the post type
I made an incorrect assumption that post_type_object only existed for Posts, but it is a reference to the wp_post table, which contains, essentially, everything.
So the code needs to be adapted so that it also works for the following types:
- Pages
- Categories
- Tags
- Attachments / Media
- Custom Post types

I'll update the PR so that it better reflects the current state.

Update code styling

1. Add comments where possible.
2. Split up the big Advanced Settins with several Functional Components.
3. Make eslint happy by removing spaces, binding functions, and adding docs.
4. Rename variables to better reflect their function.

Remove the PHPCS violation