Tests: fix test data
The array as previously defined, contained two entries with an empty string as key + value.
In effect, that meant that the array only contained one of those entries as the second empty key would overwrite the first.

Leaving the values empty, but passing keys will allow the test to actually test with two empty values in the array.