Documentation: remove file comments from namespaced classes
YoastCS 1.2.0 introduced a new YoastCS native sniff which makes it unnecessary for namespaced classes to have a file comment.

For the relevant classes to which this currently applies, I've removed the file comments and where relevant, merged the file short description and class short description to provide a more comprehensive class description.

This commit includes:
* Fixing class description and (file) descriptions to be more accurate and more consistent.
* Removing redundant `@package` tags in the class docblocks.
* Remove redundant repetition of the class name in the class docblock if there is a proper description available.
* Minor punctuation fixes.

Includes adjusting the CS thresholds:
* Lowered the number of expected warnings as a lot are fixed by this PR.
* Upped the number of expected errors as the `trunk` build was already failing on this.