Tests: remove use of `assertAttributeInstanceOf()` assertions [2]
The `assertAttribute*()` assertion methods were deprecated in PHPUnit 8.x and removed in PHPUnit 9.

The reasoning for that was that `protected` and `private` properties are implementation details and should not be tested directly, but via the methods in the class.

For now, I've elected to maintain the existing test behaviour, while making it work in a PHPUnit cross-version compatible manner.

This is done by:
1. Using a test helper method introduced provided via PHPUnit Polyfills (0.2.0) to get the value of a non-public property.
2. Changing the assertions for the tests from `assertAttribute*()` to `assert*()`
3. **Testing the class under test directly.**
    In these three cases, either a Mock or a Double of the class was used in combination with testing the value of a `private` property. This will no longer work as the private property is set in the parent class and inaccessible from the mock/double child class. Overloading the visibility of the property in the mock/double is also not a solution.
    Opting for the simplest of solutions now, by using a local instance of the original class to still test the constructor.

At a later point in time, it should be re-examined whether these implementation details should be tested in this way, if at all.