Admin\MyYoast_Proxy_Test: refactor tests using `at()`
The `at()` matcher has been deprecated in PHPUnit 9 and will be removed in PHPUnit 10. It is strongly recommended not to rely on the order in which methods are invoked.

Replacing the expectations create with the `at()` matcher, by a slightly different expectation which still tests that the `set_header()` function is called 1) the expected number of times and 2) with specific arguments and 3) in a certain order.

This stabilizes these tests for the future.