Compile-DI: fix custom loader on Windows
Both the returned value from `glob` as well as the classmap generated by Composer contain file paths with a mix of forward and backward slashes when run on Windows.

This was breaking the `compile-di` command.

Previously it was _silently_ broken - it wouldn't work, but without throwing an error.
Since the addition of the `Schema_Blocks` it would fail with an exception.

This commit fixes this by:
1. Adding a `normalize_slashes()` method.
2. Calling that method in the relevant places in the script to make sure that paths are normalized before they are compared based on their string value.