Replace the activation action
The problem with the activate save buttons is that it does not set the editor changed value to true.
The save draft checks the isChanged value. Resulting in an ignored save. To fix this we set isChanged to true.

But the `document/save/set-is-modified` does the isChanged and the `status:change` trigger for us. So using that here instead.