Composer: update Yoast/wp-test-utils to 0.2.1
A new version of the WP Test Utils package has been released which deals with the necessary work-arounds for the WP integration tests and adds utilities for this which individual plugins can use, like functions for the bootstrap and a `TestCase` which includes all the polyfills from the PHPUnit Polyfill repo.

As this repo does **not** have a `config.platform.php` set to PHP `5.6` in the `composer.json` file, as dev dependencies need a higher PHP version, the `require-dev` for PHPUnit needs to remain to keep it locked at PHPUnit 5.7.

The update to WP Test Utils `0.2.1` also updates BrainMonkey to version `2.6.0`.

Refs:
* https://github.com/Yoast/wp-test-utils/releases/tag/0.2.0
* https://github.com/Yoast/wp-test-utils/releases/tag/0.2.1