Tests: remove use of `assertAttribute*()` assertions
The `assertAttribute*()` assertion methods were deprecated in PHPUnit 8.x and removed in PHPUnit 9.

The reasoning for that was that `protected` and `private` properties are implementation details and should not be tested directly, but via the methods in the class.

For now, I've elected to maintain the existing test behaviour, while making it work in a PHPUnit cross-version compatible manner.

This is done by:
1. Using a test helper method introduced provided via PHPUnit Polyfills (0.2.0) to get the value of a non-public property.
2. Changing the assertion for the tests from `assertAttribute*()` to `assert*()`

At a later point in time, it should be re-examined whether these implementation details should be tested in this way.

Note: as these tests will only be run on PHPUnit 5-7, this change isn't _strictly_ necessary, however, it makes the use of the methods in tests consistent across the various plugins and should help devs to remember to use this pattern rather than the old `assertAttribute*()` methods.