Unit tests: minor tweak
As the WP Test Utils now brings in a `expectOutputContains()` method with a slightly different method signature, we need to update the method signature of the method with the same name in the BrainMonkey base `TestCase` class.

This is a temporary fix. In the (near) future, this method will be removed in favour of the one in WP Test Utils. Same as is already done in this PR for the integration tests.