Composer/actions: update the PHPCS threshold check for GHA compatibility
While running some tests with GitHub Actions, one of the findings was that GHA has much better process isolation than Travis.
In practice, this means that querying the `YOASTCS_ABOVE_THRESHOLD` constant as set by the YoastCS `Threshold` report is not going to work as the calling process does not have access to it.

The result of this would be that the branch report would not be run when needed and that the exit code would always be `0`, which would prevent GHA from failing the build when the violations do not comply with the set thresholds.

I've also tested to see if using an environment variable would work better, but no luck there either.
You can see various experiments I've run here: https://github.com/jrfnl/composer-script-poc/actions

The short of it is, that for the script to be compatible with both Travis, as well as GH Actions, the content of the report needs to be examined instead.